backend/requirements.txt
fastapi==0.111.0
pymongo==3.9.0
python-telegram-bot==21.3
requests==2.32.3
uvicorn==0.30.1
web3==6.10.0
